Add logout, upgrade AppGuard, upgrade httpClientService: add lock to sync all refresh request, fix bugs: refresh fail

+++ src/app/service/auth.service.ts
@@ -9,28 +9,37 @@ import { AuthRequest } from '../models/request/auth.request';
 })
 export class AuthService {
 
-  constructor(private httpClient: HttpClientService,
+  constructor(
+    private httpClient: HttpClientService,
     private cookieService: CookieService,
-    private router: Router) { }
+  ) { }
 
   auth = async (auth: AuthRequest) => await this.httpClient.post('api/auth/token', auth);
-  
-  register = async (auth: AuthRequest) => await this.httpClient.post('api/player/register', auth);
+
+  logout = async () => await this.httpClient.postWithAuth('api/auth/my-token/invalidate', {});
   
   authenticated = async () => {
     const token = this.cookieService.getToken();
-    if(!token) return false; 
-
+    // Không có access token
+    if (!token) 
+      return false; 
+    // Kiểm tra tính hợp lệ của access token
+    const res = await this.httpClient.getWithAuth('api/auth/token/introspect', {});
+    // Access token không hợp lệ
+    if (!res)
+      return false;
+    // Return 
     return true;
-    const res = await this.httpClient.get('', {token: token});
-    return res;
   }
 
-  login = () => this.router.navigate(["login"]);
-
-  logout = () => {
-    this.cookieService.deleteToken();
-    this.router.navigate(['login']);
+  handleLogout = async () => {
+    // Invalidate token request
+    const res = await this.logout();
+    // Delete tokens
+    if (res) {
+      this.cookieService.deleteToken();
+      this.cookieService.deleteRefreshToken();
+    }
+    return true;
   }
- 
-}
+ }
